[DEV] Remove unused demangle function
