+++ src/libs/antares/Enum.hxx
@@ -32,7 +32,7 @@
 #include <list>
 
 #include <antares/exception/AssertionError.hpp>
-#include <antares/stdcxx/demangle.hpp>
+#include <antares/stdcxx/class_name.h>
 
 #include <antares/Enum.hpp>
